# 3° Anno/Programmazione per il Web/Progetto/WASAText/doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000"
info:
  title: WASAText
  description: una piattaforma di messaggistica che supporta chat individuali e di gruppo, accessibile tramite PC
  version: 1.0.0
...
paths:
...
  /users/search:
    get:
      tags:
        - utente
      summary: Ricerca utenti per username
      description: Restituisce la lista di utenti il cui username corrisponde al parametro di ricerca.
      operationId: searchUsers
      parameters:
        - $ref: '#/components/parameters/UserUsername'
      responses:
        '200':
          description: Elenco utenti trovati
          content:
            application/json:
              schema:
                type: array
                description: Lista di utenti trovati
                items:
                  $ref: '#/components/schemas/User'
                minItems: 0
                maxItems: 100
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # SEZIONE 1.2 - Impostazioni nome utente
  /users/me/name:
    put:
      tags:
        - utente
      summary: Imposta o aggiorna il nome utente
      description: |-
        Permette all'utente autenticato di impostare o modificare il proprio nome.
        Restituisce l'oggetto utente aggiornato.
      operationId: setMyUserName
      security:
        - bearerAuth: []
      requestBody:
        description: Nuovo nome utente
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetNameRequest'
      responses:
        '200':
          description: Nome utente aggiornato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          $ref: '#/components/responses/Conflict'
          

  # SEZIONE 1.3 - Impostazioni foto profilo
  /users/me/photo:
    put:
      tags:
        - utente
      summary: Imposta o aggiorna la foto profilo dell'utente
      description: |-
        Permette all'utente autenticato di impostare o modificare la propria foto profilo.
        Restituisce l'oggetto utente aggiornato.
      operationId: setMyPhoto
      security:
        - bearerAuth: []
      requestBody:
        description: Nuova foto profilo
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPhotoRequest'
      responses:
        '200':
          description: Foto profilo aggiornata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
 
  # SEZIONE 1.4 - Gestione mio account utente 
  /users/me:
    get:
      tags:
        - utente
      summary: Ottieni i dettagli dell'utente autenticato
      description: Restituisce i dettagli dell'utente autenticato.
      operationId: getMyProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dettagli utente recuperati con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - utente
      summary: Elimina l'account dell'utente autenticato
      description: |-
        Permette all'utente autenticato di eliminare il proprio account.
      operationId: deleteMyAccount
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Account eliminato con successo
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # SEZIONE 2 - Le Conversazioni (Chat)
...
  /chats/{chat_id}/messages/{message_id}/reaction: 
    parameters:
        - name: chat_id
          in: path
          description: ID univoco della chat.
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
            pattern: '^[a-zA-Z0-9_-]{1,64}$'
        - name: message_id
          in: path
          description: ID univoco del messaggio.
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
            pattern: '^[a-zA-Z0-9_-]{1,64}$'
    post: 
      tags:
        - reazioni 
      summary: Inserimento di una reazione ad un messaggio
      description: |-
        Permette di inserire una reazione (o meglio una emoji) in un messaggio specifico 
        all'interno della chat 
      operationId: commentMessage
      security:
        - bearerAuth: []
      requestBody:
        description: Dettagli della reazione da inserire
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReactionRequest'
      responses:
        '201':
          description: Reazione inserita con successo 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - reazioni 
      summary: Rimozione di una reazione ad un messaggio
      description: |-
        Permette di rimuovere una reazione (o meglio una emoji) in un messaggio specifico 
        all'interno della chat 
      operationId: uncommentMessage
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Reazione rimossa con successo
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # SEZIONE 3.4 - Inoltro di uno o più messaggi ad un'altra chat
...
components:
...
  schemas:
...
    User:
      type: object
      description: Rappresenta un utente nella piattaforma WASAText
      required:
        - name
      properties:
        id:
          type: string
          minLength: 6
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]{6,64}$'
          example: "abcdef012345"
          description: Identificatore univoco dell'utente
        name:
          type: string
          pattern: '^[a-zA-Z0-9._-]{1,64}$'
          example: "Maria"
          minLength: 1
          maxLength: 64
          description: Nome utente univoco
        photo_url:
          type: string
          format: uri
          description: URL della foto profilo dell'utente
          example: "http://example.com/photos/user123.png"
          minLength: 1
          maxLength: 2048
          pattern: '^https?://.+$'
...
    Reaction:
      type: object 
      description: Rappresenta una reazione a un messaggio tramite emoji
      required:
        - id
        - message_id
        - user_id
        - emoji_code
        - created_at
      properties:
        id:
          type: string
          minLength: 6
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]{6,64}$'
          example: reaction1
          description: Identificatore univoco della reazione
        message_id:
          type: string
          minLength: 6
          maxLength: 64 
          pattern: '^[a-zA-Z0-9_-]{6,64}$'
          example: message1234
          description: Identificatore del messaggio a cui è associata la reazione
        user_id: 
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]{1,64}$'
          example: user1
          description: Identificatore dell'utente che ha inserito la reazione
        user_name:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9._-]{1,64}$'
          example: Maria
          description: Nome dell'utente che ha inserito la reazione
        emoji_code:
          type: string
          minLength: 1
          maxLength: 4 
          description: Codice dell'emoji usata come reazione
          pattern: '^[\u0000-\uFFFF]{1,4}$'
          example: "👍🏻"
        created_at:
          type: string 
          format: date-time
          example: "2025-10-30T14:48:00.000Z"
          description: Data e ora di creazione della reazione

    # Serve anche uno schema di input per le reazioni
    CreateReactionRequest:
      type: object
      description: Rappresenta la richiesta di creazione di una reazione a un messaggio tramite emoji
      required:
        - emoji_code
      properties:
        emoji_code:
          type: string
          minLength: 1
          maxLength: 4
          pattern: '^[\u0000-\uFFFF]{1,4}$'
          example: "👍🏻"
          description: Codice dell'emoji usata come reazione

    # Creo gli schemi per le richieste dei vari endpoint
    SetNameRequest:
      type: object
      description: Richiesta per impostare o aggiornare il nome utente
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 16
          description: Nuovo nome utente
          pattern: '^[a-zA-Z0-9._-]{3,16}$'
          example: "NuovoNomeUtente"
    SetPhotoRequest:
      type: object
      required:
        - photo_url
      description: Richiesta per impostare o aggiornare la foto profilo
      properties:
        photo_url:
          type: string
          format: uri
          description: URL della nuova foto profilo
          example: "http://example.com/photos/newphoto.png"
          minLength: 1
          maxLength: 500
          pattern: '^https?://.+$'
...
    Error:
      type: object
      description: Rappresenta un errore generico restituito dalle API
      properties:
        code:
          type: string
          minLength: 3
          maxLength: 64
          pattern: '^[A-Z_]{3,64}$'
          example: "NOT_FOUND"
          description: Codice dell'errore
        message:
          type: string
          minLength: 3
          maxLength: 64
          pattern: '^[\s\S]{3,64}$'
          example: "La risorsa richiesta non è stata trovata."
          description: Descrizione dettagliata dell'errore

  # RESPONSES

  # Risposte per gli errori
  responses:
    BadRequest:
      description: Richiesta non valida (400 Bad Request)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Risorsa non trovata (404 Not Found)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflitto nella richiesta (409 Conflict)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Errore interno del server (500 Internal Server Error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # PARAMETERS
  parameters:
    UserUsername:
      name: username
      in: query
      description: Username univoco dell'utente.
      required: true
      schema:
        type: string
        pattern: '^[a-zA-Z0-9._-]{1,64}$'
        minLength: 1
        maxLength: 64
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT